+++ search/src/optimizer_generics.rs
@@ -3,8 +3,8 @@
 pub trait Optimizer {
     const IS_MAXIMIZER: bool;
     type Opposite: Optimizer;
-    fn compare(old: f32, new: f32) -> bool;
-    fn compare_for_assign(old: f32, new: f32) -> f32;
+    fn compare(old: i32, new: i32) -> bool;
+    fn compare_for_assign(old: i32, new: i32) -> i32;
 }
 
 
@@ -17,11 +17,11 @@ impl Optimizer for Maximizer {
     const IS_MAXIMIZER: bool = true;
     type Opposite = Minimizer;
     #[inline(always)]
-    fn compare(old: f32, new: f32) -> bool {
+    fn compare(old: i32, new: i32) -> bool {
         new > old
     }
     #[inline(always)]
-    fn compare_for_assign(old: f32, new: f32) -> f32 {
+    fn compare_for_assign(old: i32, new: i32) -> i32 {
         old.max(new)
     }
 }
@@ -31,11 +31,11 @@ impl Optimizer for Minimizer {
     const IS_MAXIMIZER: bool = false;
     type Opposite = Maximizer;
     #[inline(always)]
-    fn compare(old: f32, new: f32) -> bool {
+    fn compare(old: i32, new: i32) -> bool {
         new < old
     }
     #[inline(always)]
-    fn compare_for_assign(old: f32, new: f32) -> f32 {
+    fn compare_for_assign(old: i32, new: i32) -> i32 {
         old.min(new)
     }
 }
